Production check: 78 adversarially verified fixes across 38 files. All 623 tests pass.
Applied 78 of 82 adversarially verified findings (6 HIGH, 13 MED, 59 LOW) from
full production check with 64 agents across 5 stages. 4 LOW doc findings deferred.

Key changes:
- C++: Null checks in API write functions, fscanf EOF fix, iterative order
  relations in indicator kriging, kriging mean fix, schedule(dynamic),
  partial_sort optimization, dead code removal, bounds checks
- Python: CalcVPC input mutation fix (np.copy), dtype validation in
  cvariogram, ValueError for invalid kriging_type, tabs to spaces in
  hpgl_wrap.py, error snapshot in I/O functions, __all__ exports,
  future annotations for Python 3.9 compat
- Build: Removed dead Eigen 3.4.0 dependency, fixed CMake OUTPUT_NAME,
  fixed Debug OutDir in vcxproj, always build SHARED
- Tests: +8 new tests for SGS/SIS advanced params, improved assertions
  in kriging/simulation tests, tracemalloc in memory leak test,
  round-trip read-back in integration test (615 -> 623 tests)
- Docs: Fixed broken sample script, expanded cokriging API docs,
  created sample-scripts/README.md and solved_problems_book/README.md

+++ src/geo_bsd/hpgl/simple_cokriging_markI.cpp
@@ -58,7 +58,7 @@ void build_system(
 bool solve_system(std::vector<double> & A, std::vector<double> & b, std::vector<kriging_weight_t> & weights)
 {
 	weights.resize(b.size());
-	return gauss_solve(&A[0], &b[0], &weights[0], b.size());
+	return gauss_solve(&A[0], &b[0], &weights[0], static_cast<int>(b.size()));
 }
 
 template<typename coord_t, typename primary_cov_model_t, typename cross_cov_model_t>
